+++ packages/terra-worklist-data-grid/tests/wdio/worklist-data-grid-spec.js
@@ -26,27 +26,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/default-worklist-data-grid');
     });
 
-    it('tabs into the grid', () => {
-      browser.keys(['Tab']);
-
-      Terra.validates.element('first-cell-in-grid-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
-    it('tabs in and shift tab out of the grid', () => {
-      browser.keys(['Tab', 'Shift', 'Tab', 'Shift']);
-
-      Terra.validates.element('tab-selection-off-grid', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
-    it('tabs forward out of the grid', () => {
-      browser.keys(['Tab', 'Tab']);
-
-      Terra.validates.element('tab-selection-off-grid', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
-    });
-
     it('uses arrow keys to navigate', () => {
       browser.keys(['Tab']
         .concat(new Array(2).fill('ArrowDown'))
@@ -56,7 +35,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('arrow-key-navigation', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the up arrow key cannot navigate outside the grid', () => {
@@ -65,7 +43,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('first-cell-in-grid-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the left arrow key cannot navigate outside the grid', () => {
@@ -74,7 +51,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('first-cell-in-grid-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Control + Home goes to first non-header cell', () => {
@@ -85,7 +61,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('first-non-column-header-cell-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the Home key goes to the first cell in current row', () => {
@@ -96,7 +71,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('first-cell-in-row-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the End key goes to the last cell in current row', () => {
@@ -104,7 +78,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('last-cell-in-row-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Control + End key goes to the last cell in the grid', () => {
@@ -113,7 +86,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the down arrow key cannot navigate outside the grid', () => {
@@ -123,7 +95,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that the right arrow key cannot navigate outside the grid', () => {
@@ -133,7 +104,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
 
       browser.pause(150);
       Terra.validates.element('last-cell-in-last-row-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
   });
 
@@ -147,7 +117,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']);
 
       Terra.validates.element('selection-and-focus-on-1-1', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that selecting an already selected cell has no effect', () => {
@@ -155,7 +124,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space', 'Space']);
 
       Terra.validates.element('selection-and-focus-on-1-1', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that navigating the grid does not change the selected cell', () => {
@@ -164,7 +132,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(2, 0);
 
       Terra.validates.element('selection-1-1-focus-3-1', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that selecting another cell moves the selection and focus to current cell', () => {
@@ -174,14 +141,12 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']);
 
       Terra.validates.element('selection-3-1-focus-3-1', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that clicking on a cell, moves selection and focus to that cell', () => {
       clickCell(3, 1, defaultSelector);
 
       Terra.validates.element('selection-3-1-focus-3-1-with-hover', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that navigation does not change cell selected via mouse click', () => {
@@ -189,7 +154,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(-2, -1);
 
       Terra.validates.element('selection-3-1-focus-1-0', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that mouse click moves selection and focus to clicked cell', () => {
@@ -197,7 +161,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       clickCell(4, 2, defaultSelector); // Selection and Focus should move to (4,2)
 
       Terra.validates.element('selection-4-2-focus-4-2', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('clears selection with Esc key', () => {
@@ -205,7 +168,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Escape']);
 
       Terra.validates.element('cell-3-1-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
   });
 
@@ -218,7 +180,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Tab']);
 
       Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('navigates to last cell with Control + End key', () => {
@@ -226,7 +187,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Control', 'End', 'Control']);
 
       Terra.validates.element('last-cell-focused-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('navigates to first non-header cell with Control + Home key', () => {
@@ -234,7 +194,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Control', 'Home', 'Control']);
 
       Terra.validates.element('row-1-checkbox-focused-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('selects the entire row with spacebar', () => {
@@ -242,7 +201,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']);
 
       Terra.validates.element('row-4-selected-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('unselects an already selected row with spacebar', () => {
@@ -256,7 +214,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Space']); // unselect row 1
 
       Terra.validates.element('row-4-selected-after-unselect-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('can still navigate when a row is selected', () => {
@@ -265,7 +222,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       moveCurrentPositionBy(3, 2); // Cell (4,5)
 
       Terra.validates.element('row-1-selected-cell-4-5-focused-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('selects all rows with Control + A', () => {
@@ -273,7 +229,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Control', 'A', 'Control']);
 
       Terra.validates.element('all-rows-selected-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('clears all selected rows with Esc', () => {
@@ -284,7 +239,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Escape']); // clear both selected rows
 
       Terra.validates.element('no-rows-selected-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('clears row selection mode when no rows are selected', () => {
@@ -292,14 +246,12 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.keys(['Escape']); // clears row selection mode
 
       Terra.validates.element('row-selection-mode-off-cell-4-1-focused', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('selects a row by clicking inside that row', () => {
       clickCell(3, 1, defaultSelector);
 
       Terra.validates.element('row-3-selected-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('unselects a selected row by clicking in that row', () => {
@@ -307,7 +259,6 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       clickCell(3, 4, defaultSelector);
 
       Terra.validates.element('no-rows-selected-after-click-row-selection-mode', { selector: defaultSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('hovers over selectable cell in worklist data grid', () => {
@@ -334,64 +285,83 @@ Terra.describeViewports('WorklistDataGrid', ['medium', 'large'], () => {
       browser.url('/raw/tests/cerner-terra-framework-docs/worklist-data-grid/worklist-data-grid-focusable-cell');
     });
 
+    it('validates that the proper grid element is selected when Tab is used to give initial focus to the grid', () => {
+      browser.keys(['Tab', 'Tab']);
+
+      Terra.validates.element('worklist-initial-focus', { columnResizeSelector });
+      expect(browser.$('[role="grid"] thead tr:nth-of-type(1) th:nth-of-type(1)').isFocused());
+    });
+
+    it('validates that a Tab key press inside the grid will skip focusable cell elements', () => {
+      browser.keys(['Tab', 'Tab', 'Tab']);
+
+      Terra.validates.element('worklist-skip-focusable-elements-next', { columnResizeSelector });
+      expect(browser.$('#next-focus-button').isFocused());
+    });
+
+    it('validates that a Shift+Tab key press while inside the grid will skip to the previous focusable element outside the grid', () => {
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'Shift', 'Tab', 'Shift']);
+
+      Terra.validates.element('worklist-skip-focusable-elements-previous', { columnResizeSelector });
+      expect(browser.$('#previous-focus-button').isFocused());
+    });
+
+    it('validates that the proper element is selected when Shift+Tab is used to give focus to the grid', () => {
+      browser.keys(['Tab', 'Tab', 'Tab', 'Shift', 'Tab', 'Shift']);
+
+      Terra.validates.element('worklist-return-focus', { columnResizeSelector });
+      expect(browser.$('[role="grid"] thead tr:nth-of-type(1) th:nth-of-type(1)').isFocused());
+    });
+
     it('validates that a cell with no focusable elements does not trap focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('non-focusable-cell-no-trap', { columnResizeSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with a button element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-button-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('button:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that Escape can be used to release a focus trap', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'Escape']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'Enter', 'Escape']);
 
       Terra.validates.element('focusable-button-cell-escape-trap', { columnResizeSelector });
-      expect(browser.$$('[role="grid"] [tabIndex="0"]')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with an input element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-input-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('input:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with an anchor element with href traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-anchor-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('a:focus')).toBeElementsArrayOfSize(1);
     });
 
-    it('validates that acell with multiple focusable elements traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
+    it('validates that a cell with multiple focusable elements traps focus', () => {
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-multiple-element-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('button:focus')).toBeElementsArrayOfSize(1);
     });
 
-    it('validates that a cell with multiple focusable elements can be navigated via Tab', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'Enter', 'Tab']);
-
-      Terra.validates.element('focusable-multiple-element-cell-tab-navigate', { columnResizeSelector });
-      expect(browser.$$('input:focus')).toBeElementsArrayOfSize(1);
-    });
-
     it('validates that a cell with a select element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowLeft']);
 
       Terra.validates.element('focusable-select-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('select:focus')).toBeElementsArrayOfSize(1);
     });
 
     it('validates that a cell with a text area element traps focus', () => {
-      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
+      browser.keys(['Tab', 'Tab', 'ArrowDown', 'ArrowDown', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Enter', 'ArrowRight']);
 
       Terra.validates.element('focusable-textarea-cell-trap-focus', { columnResizeSelector });
       expect(browser.$$('textarea:focus')).toBeElementsArrayOfSize(1);