+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Updated focusable cell example for `terra-worklist-data-grid` to allow text area to be editable.
 
 ## 1.32.0 - (August 10, 2023)
 
@@ -27,6 +29,9 @@
 
 * Added
   * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.
+
+* Changed
+  * Updated tests for `terra-worklist-data-grid` to verify focusable cell elements do not receive focus on tab key press.
 
 ## 1.30.0 - (August 2, 2023)